Add kubejobs progress plugin
KubeJobs is a plugin that watch the number of tasks finished by a
k8s Job resource. It compares with the total number of tasks
(considering the workload size) and with the proportional time
spent based on a deadline. The difference between time and tasks
progresses is called error and published in a redis queue service
and also in Monasca (if graphic_metrics parameter is set True).